Initial Separation Stage Added Seminar and Problem app to settings.py Moved all url mappings, models, views, admin page mappings and template files to separate app
