+++ src/plus_study/urls.py
@@ -18,5 +18,7 @@ from django.contrib import admin
 
 urlpatterns = [
     url(r'^admin/', admin.site.urls),
+    url(r'^seminar/', include('seminar.urls')),
+    url(r'^problem/', include('problem.urls')),
     url(r'^', include('website.urls')),
 ]
